perf(utxo/postgres): hot-path diet — pack per-output arrays, INT4 height/vout columns
Array packing (the headline): the 5 per-output parallel array columns
(utxo_hashes BYTEA[], out_spendables/out_frozens BOOLEAN[],
coinbase_spending_heights BIGINT[], spendable_ins INT[]) become a flat
32-byte-stride utxo_hashes BYTEA + get_bit-compatible bitmaps + out_count/
spendable_count scalars + a single coinbase_spending_height (the per-output
array was redundant — every output of a coinbase shares one maturity height).
This deletes the out_arr array_agg GROUP BY re-aggregation CTE from the bulk
create path (a top-4 server-CPU statement) and the per-output param fan-out;
per-output access in the spend validation CTEs is substr()/get_bit() byte
arithmetic. Encoding agreement between the Go packers and SQL
substr()/get_bit() is locked in by packed_outputs_test.go against live
postgres. The deferred-DAH fully-spent comparand reads the spendable_count
scalar instead of array_positions().

INT4 narrowing: spends.prev_output_idx + spent_at_height and the txs height
columns (unmined_since, delete_at_height, preserve_until, mined_at_height,
coinbase_spending_height) BIGINT -> INT, with explicit vout overflow guards
where uint32 enters. Spends UNIQUE-index entries -13%, heap rows -10%.

Measured (M3 Max, 10K workers, sustained-with-prune, table bounded, honesty
gate quiet): 2-shard 92.4K -> 95.3/95.7K median (CV 2.9-3.9%) from packing;
INT4 neutral-within-drift on this CPU-bound box (87.9/92.8/94.4K across three
runs, late reps touching 100.5K) — kept for the production cache-footprint
win (narrower rows = more of the working set resident). Two burst-shaped
knobs were A/B'd and REJECTED with rationale comments left in place: 2ms
batcher ticks (95.7K -> 88K) and minedChunkSize=4000 (95.7K -> 86K) — both
break the create/reclaim interleaving.

Schema is not backward compatible with v7-array databases (fresh-schema
deployments only; migration note in schema.go). Full postgres suite green
incl. -race; oracle invariants unchanged.

+++ stores/utxo/postgres/iterators.go
@@ -50,7 +50,7 @@ func newPrunableUnminedTxIterator(store *Store, cutoffBlockHeight uint32) (*unmi
 		ORDER BY hash
 	`
 
-	rows, err := store.pool.Query(context.Background(), q, int64(cutoffBlockHeight))
+	rows, err := store.pool.Query(context.Background(), q, int32(cutoffBlockHeight)) // unmined_since is INT4
 	if err != nil {
 		return nil, err
 	}
@@ -245,7 +245,7 @@ func (s *Store) QueryOldUnminedTransactions(ctx context.Context, cutoffBlockHeig
 		SELECT hash FROM txs
 		WHERE unmined_since IS NOT NULL AND unmined_since <= $1
 		ORDER BY unmined_since LIMIT 1000
-	`, int64(cutoffBlockHeight))
+	`, int32(cutoffBlockHeight))
 	if err != nil {
 		return nil, errors.NewStorageError("failed to query old unmined transactions", err)
 	}
