+++ stores/utxo/postgres/preservation.go
@@ -32,7 +32,7 @@ func (s *Store) PreserveTransactions(ctx context.Context, txIDs []chainhash.Hash
 		inClause, args := buildINClauseLocal(chunk, 2)
 		q := fmt.Sprintf(`UPDATE txs SET preserve_until = $1, delete_at_height = NULL WHERE hash IN %s`, inClause)
 
-		allArgs := append([]interface{}{int64(preserveUntilHeight)}, args...)
+		allArgs := append([]interface{}{int32(preserveUntilHeight)}, args...) // preserve_until is INT4
 		result, err := s.pool.Exec(ctx, q, allArgs...)
 		if err != nil {
 			return errors.NewStorageError("[PreserveTransactions] failed to preserve chunk: %v", err)
@@ -49,15 +49,16 @@ func (s *Store) PreserveTransactions(ctx context.Context, txIDs []chainhash.Hash
 // ProcessExpiredPreservations handles transactions whose preservation period has expired.
 // Sets delete_at_height = currentHeight + retention and clears preserve_until.
 func (s *Store) ProcessExpiredPreservations(ctx context.Context, currentHeight uint32) error {
-	// Widen to int64 before adding so the sum cannot wrap in uint32 arithmetic.
+	// Widen to int64 before adding so the sum cannot wrap in uint32 arithmetic,
+	// then narrow to int32 for the INT4 delete_at_height column.
 	// Stamping on expiry is unconditional by design (matches the aerospike store:
 	// the pruner's safety checks are the gate, not stamp-time eligibility).
-	deleteAtHeight := int64(currentHeight) + int64(s.settings.GetUtxoStoreBlockHeightRetention())
+	deleteAtHeight := int32(int64(currentHeight) + int64(s.settings.GetUtxoStoreBlockHeightRetention()))
 
 	result, err := s.pool.Exec(ctx, `
 		UPDATE txs SET delete_at_height = $1, preserve_until = NULL
 		WHERE preserve_until IS NOT NULL AND preserve_until <= $2
-	`, deleteAtHeight, int64(currentHeight))
+	`, deleteAtHeight, int32(currentHeight))
 	if err != nil {
 		return errors.NewStorageError("failed to process expired preservations", err)
 	}
